requirements.txt
supabase>=2.0.0
python-dotenv==1.0.0
click==8.1.7
psycopg2-binary==2.9.9
requests==2.31.0
tqdm==4.66.1
tabulate==0.9.0
